digest_directives: rename file to digest_types
